+++ Opcodes/dcblockr.c
@@ -183,8 +183,8 @@ static int32_t dcblock2(CSOUND *csound, DCBlock2* p)
       }
       out[i] = (MYFLT)(del - x1);
 
-      p1 = (p1 == del1size-1 ? 0 : p1 + 1);
-      p2 = (p2 == iirdelsize-1 ? 0 : p2 + 1);
+      p1 = ((uint64_t) p1 == del1size-1 ? 0 : p1 + 1);
+      p2 = ((uint64_t) p2 == iirdelsize-1 ? 0 : p2 + 1);
     }
 
     p->dp1 = p1; p->dp2 = p2;
